switched rodio to kira

+++ src/config.rs
@@ -12,7 +12,7 @@ pub enum Command {
     PlaySound {
         button: usize,
         file: String,
-        volume: f32,
+        volume: f64,
     },
     ToggleAudioMute {
         button: usize,
